+++ metadata-ingestion-modules/dagster-plugin/tox.ini
@@ -1,0 +1,35 @@
+# tox (https://tox.readthedocs.io/) is a tool for running tests
+# in multiple virtualenvs. This configuration file will run the
+# test suite on all supported python versions. To use it, "pip install tox"
+# and then run "tox" from this directory.
+
+[tox]
+envlist = py3-quick,py3-full
+
+[gh-actions]
+python =
+    3.6: py3-full
+    3.9: py3-full
+
+# Providing optional features that add dependencies from setup.py as deps here 
+# allows tox to recreate testenv when new dependencies are added to setup.py.
+# Previous approach of using the tox global setting extras is not recommended  
+# as extras is only called when the testenv is created for the first time!
+# see more here -> https://github.com/tox-dev/tox/issues/1105#issuecomment-448596282
+
+[testenv]
+deps = 
+    -e ../../metadata-ingestion/[.dev]
+commands =
+    pytest --cov={envsitepackagesdir}/datahub --cov={envsitepackagesdir}/datahub_provider \
+        py3-quick: -m 'not integration and not slow_integration' --junit-xml=junit.quick.xml \
+        py3-full: --cov-fail-under 65 --junit-xml=junit.full.xml \
+        --continue-on-collection-errors \
+        -vv
+
+setenv =
+    DAGSTER_HOME = /tmp/dagster/thisshouldnotexist-{envname}
+
+[testenv:py3-full]
+deps =
+    ../../metadata-ingestion/.[dev]
